XFAIL a test following the +nan fix.
This test is expecting "nan", while the fix enh made to libc makes it
now be "+nan". I have a feeling this is really a bug with either the
test or Darwin/the BSDs, since we found this was actually a bug in the
BSD libc, but just XFAIL it for us for now until I diagnose properly.

Change-Id: Icf5c7f7a0fdd9a8696842201bee8464a11e7c6f1