Add apexer the tool for creating APEX files
apexer is the tool for creating APEX files which are new package format
for system modules.

It packages a set of files under the input directory into a filesystem
image (ext4) which is then embedded to a zip file uncompressed. The
packaged files are accessible by mounting the embedded filesystem image
using the loopback device.

Bug: 112458021
Test: runtests.sh which creates an apex file from random contents and
ensures that the files inside the apex are identical to the input

Change-Id: I36a84a9dbf6b446bf23ea59110a225260cb4bbd1