Set uid/gid/mode of files in an APEX
apexer now supports --canned_fs_config to set uid, gid, mode and
capability of files in the APEX.

The syntax of a canned_fs_config is as simple as

path uid gid mode [capability]

for each path.

Currently, uid, gid, and mode should be numeric, e.g., 100 200 0644,
symbolic names are not supported yet.

Note that when --canned_fs_config is specified, it must cover ALL
files in the APEX. Missing a file is an error.

Bug: 112458021
Test: runtests.sh
Change-Id: I73da91dfd0be2a12acc80d7e08127994fa49e073