Simplify apexRule for copying files to the image directory
Previously, for copying files to the image directory, several shell
utilities such as tr, ack, dirname, xargs, etc. were used to create the
copy commands from the list of <src>:<dest_in_apex>.

Simplify this by creating the commands in soong so that we don't need to
rely on the shell utilities whose behavior and syntax are slightly
different across Linux and Mac.

Bug: 112672359
Test: m apex.test

Change-Id: Ia955af178a65c71a53260f10c21419dff30914ec