Ensure that tools that apexer is depending on are built
This fixes the problem that the internal tools that apexer is using
aren't built. They were not built because dependencies to them were
specified using 'required' property of the apexer. Since Soong does not
add dependency for the required modules, but just emits the information
to the make world, the tools were not built when the make world isn't
involved (e.g. when m <module> is invoked).

This fixes the problem by listing the depenencies to the build rule for
APEXs so that the tools are built before apexer is invoked by Soong.

Test: mma under /system/apex/build with clean out dir.
Change-Id: I3203b10400ea3d3bf52a96f2ce38979e3372f4ff