Update apex.test using multilib
The apex is built with compile_multilib: "both" to have both 32 and
64-bit versions of native_shared_libraries. However, at the same time,
the apex can have 32-bit or 64-bit only binaries via
multilib.* properties.

Test: m apex.test on ...
1) aosp_arm64 without TARGET_PREFER_32_BIT_EXECUTABLES=true
2) aosp_arm64 with TARGET_PREFER_32_BIT_EXECUTABLES=true
3) aosp_arm
4) uml (x86_64 only)
in all cases, vold, surfaceflinger and dex2oat are all intalled under
./bin directory of the APEX. And native libraries are installed under
both ./lib and ./lib64 directories in the case of 1) and 2).

In case of 4), drmserver, which is 32-bit only, isn't installed.

Change-Id: Ibd5645f19dfd80c7a1f798c31ce84667e23c298d