Pick the right test file
apex_e2e_tests in some cases pick the wrong test file when there are
other tests built under out/. This was happening because the test tries
to 'find' a file named test.apex under out/host/linunx-x86/testcases. If
there are other tests having a file with the same name, that file could
be choosen instead.

Fixing the problem by doing the search from the directory that is local
to apex_e2e_tests.

Also making sure that the returned package name is correctly matched
with the expected name. Previously, it just checked whether the returned
name starts with the expected name.

Bug: 120821760
Bug: 119569101
Test: atest apex_e2e_tests
Change-Id: Iaa321da04efd448ba28d45f3386fa93a308c9a07