Add --pubkey option to bundle public key with APEX
The public key file is embedded in the zip container of the APEX with
the name 'apex_pubkey'. In debuggable builds, the bundled public key
will be used as a fallback when no matching public key is found in the
built-in partitions (e.g. /system/etc/security/apex).

This is useful for testing-purpose APEXes; the public key for the APEXes
do not need to be force installed on the device, which requires rooting.

Bug: 122047804
Test: add 'installable: false' to the apex_key
'com.android.apex.test_package.key'. mma under
/system/apex/apexd/apexd_testdata. unzip the built APEX. The public key
(renamed to apex_pubkey) is found in the zip container.

Change-Id: I07be3be8ae763c12dc0541fec2b172b5f11bbafd