Quick workaround for APEX mounting failures.
apexservice_test regularly fails to mount APEX images on sailfish/marlin
and some emulators. It turns out this happens because of a race
condition, where the kernel has just created the verity device, but
userspace still needs to create the device-node entry in /dev. For now
we can work around this by retrying the mount.

Bug: 122059364
Test: apexservice_test run for 100+ times successfully
Change-Id: I20fcda365f21ee4e25b5b496e3c72a028e81ade3