Remove TEST_MAPPING from AOSP tree.
Tests require some framework changes that are internal only until next
release, in order to run properly. Remove the TEST_MAPPING file so that
no pre-submit/post-submit tests are run on AOSP tree. This file will be
submitted in the internal tree instead, and released together with the
next framework changes.

Bug: 118865310
Test: n/a
Change-Id: Ice245d5a0b4183fb28752e19910d7b7e165a1988