Listen to ro.init_updatable_crashing prop in apexd
As an additional signal to apexd update status, we listen for frequent process
crashes in init, when updatable processes crash too frequently, init sets
the ro.init_updatable_crashing property.

In this cl, apexd listens for the propery change and calls a rollback function
to unstage newly staged packages. This rollback function is just a stub at the
moment. If no packages crash and we receive the sys.boot_completed success
signal, we stop listening for process crashes from init and add the main apexd
thread to the binder thread pool.

Bug: 120598832
Test: Manually tested with logs that sys.boot_completed property change
is received in apexd_prop

Change-Id: Ic85a77e2ef634e7f5af1d90cacefd101ff55510b