submitStagedSession returns a boolean.
If verification fails, we don't want to throw a Binder exception which
in turns gets converted to a Java runtime exception, since this might be
not noticed by the caller. Return a boolean instead, and write the
ApexInfo array as an out parameter in a new parcelable called
ApexInfoList. (Lists of parcelables are currently not supported as out
parameter by native binder)

Bug: 118865310
Test: m apexservice_test
Change-Id: Ie3976afddaf7e037186a76cc821adb6cc7060a94