Apexd: Add /dev/loop open loop
apexservice_test regularly fails to mount APEX images on sailfish/marlin
and some emulators. It turns out this happens because of a race
condition, where the kernel has just created the loop device, but
userspace still needs to create the device-node entry in /dev. For now
we can work around this by retrying the open call.

Bug: 122059364
Test: Device boots
Test: atest apexservice_test
Change-Id: I6868124f871d20d792476258f1b8dcfe617a3ccf