Wait for pre-reboot verification.
This is a workaround to avoid rebooting the device when the pre-reboot
verification is not yet completed. It should be substituted with waiting
for the SESSION_UPDATE broadcast message and verify that the state is
isReady, or by checking getStagedSession.

Also, make the accessory apex test package not installable so it's not
copied under /system/apex.

Bug: 122882120
Test: atest apex_e2e_tests
Change-Id: I1503aa4f8da8c264f85b79d7ea63290aa1ff6ff4