Bootstrap apexd waits for /dev/loop-control
Becaue the bootstrap apexd is started very early - just after ueventd is
started, but before all dev nodes are confirmed to be created  -
/dev/loop-control is not guaranteed to exist when the bootstrap apexd
first runs. Our solution is to let apexd to wait for the dev node to
appear by ueventd.

Alternative considered: we could execute the bootstrap apexd after the
initialization by ueventd is finished. However, this will increase the
booting time a lot because the entire uevent initialization takes up to
500 ms. Compared to this, /dev/loop-control becomes available only 150ms
from the start of ueventd. So, by starting the bootstrap apexd early
with ueventd, about 350 ms of time is spent both for ueventd and apexd.

Bug: 125549215
Test: m device boots
Change-Id: I2c668bda6908e6ace80eb7175759f964bc795abc