Restart on dm-verity corruption.
This is consistent with what we do on /system; it has the additional
benefit of automatically rolling back a staged install in case the
device supports filesystem checkpointing.

The only time we don't use this mode is when we're verifying the
filesystem image manually, in which case we detect the read error
and abort the staged install.

Bug: 126514108
Test: With a manually corrupted apex file inspect kernel log:
[   54.670735] device-mapper: verity: 7:0: data block 4 is corrupted
[   55.681668] reboot: Restarting system with command 'dm-verity device
corrupted'

Change-Id: Iad97949c8018b3f645f16a37449cbb108956f4ab