Ensure APEXes have minSdkVersion specified
Missing minSdkVersion causes a problem when an APEX is re-signed with
sign_target_files_apks. Specifically, if the value is missing, SignApk
uses the most conservative algorithm SHA1withRSA which is being
deprecated. To fix the problem, when an APEX is built, the default
minSdkVersion is specified to ensure that all APEXes have minSdkVersion
set.

Bug: 131128233
Test: m com.android.runtime.debug; aapt dump badging <path_to_the_apex>
shows minSdkVersion

Change-Id: Ibc44b094052ea7a8ed96f9f464c55d48af37b04e