Move starting of filesystem checkpoint to apexd.
Starting a checkpoint can fail, and in that case we'd like to backup
active APEXes, in order to revert back to the previous state. Starting a
checkpoint was previously done from framework, but the logic and
interface becomes cleaner if it's done in apexd instead (where it
arguably belongs anyway).

Bug: 126740531
Test: atest apexservice_test
      atest apex_e2e_tests

Change-Id: I9b1297c1664c2d2b2f838bebdefabb9fc9e72b51