Don't use apexd when TARGET_FLATTEN_APEX == true
When TARGET_FLATTEN_APEX is true, apexd is not used to activate the
built-in flattened APEXes. Init simply bind-mounts /system/apex to
/apex.

When apexd is invoked for the flattened APEX (ro.apex.updatable is not
set or set to false), apexd just returns immediately without doing
anything but marking apexd as ready to let init to proceed without
waiting apexd.

Also, the 'critical' attribute is removed from apexd service, since
now apexd is now expected to exit. Otherwise, init will detect the exit
of apexd as crash and will refuse to boot.

Bug: 132413565
Test: marlin is bootable
Change-Id: I7d1e45b6d4c6d6594ce10817f5c5f139f10ac27f