apexd stops itself with flattened APEXes
With flattened APEXes, apexd does nothing but exits itself. However,
since the service is not a oneshot service, it is restarted forever,
causing logspam. To solve this problem, apexd stops itself in that case.
Since the service is stopped, init does not restart the service again.

Also, added 'disabled' to the apexd.rc to not start the service with the core
class. It should always be explicitly started by its name.

Bug: 133907211
Test: launch sdk_phone_x86_64
adb logcat -d | grep apexd | wc -l
returns 3

Change-Id: I88113e0a1dff018baacc29dde7c870cf759e6f03