irqchip/sunxi-nmi: support wake up
On Allwinner boards, the NMI of the SoC is usually hooked to X-Powers
AXP series PMIC, which have power key support.

As the power key of the PMIC is potentially a wake up source of the
boards, the NMI controller should be capable of wake up.

Add wake up support to NMI controller thus the AXP can wake up the
board.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>