net: stmmac: sun8i: Fix reboot issue with ethernet on Orange Pi 3
We disable both phy regulators before reboot, so that they are in a
known state on the next boot. This should be done in u-boot, but
it doesn't support AXP806 yet, so we shut it down in kernel.

Signed-off-by: Ondrej Jirman <megous@megous.com>
(cherry picked from megous commit b0a1f7a24b1bcb20720a6673a0247bcfe2591a23)
Signed-off-by: Roman Stratiienko <r.stratiienko@gmail.com>