Use PushFilePreparer in apex-rollback-tests
Calling remountSystemWritable to install test files on the system
partition leaves the device in a state that could adversely effect
subsequent tests run on the device. Switch to the more standard
PushFilePreparer to set up the device rather than have to duplicate
logic to restore the device to a good state after running the test.

Bug: 140014856
Test: atest apex_rollback_tests
Change-Id: I5d320d519d304485295613e297fc21cf470f0cba