apexd: Remove unnecessary checks on ro.apex.updatable property
If ro.apex.updatable is false, apexd won't be running on a device,
hence there is no need in checking for it in the code.

Also changed apexd_main a little bit, so that one-of invocations of
apexd (e.g. apexd --preinstall, apexd --bootstrap) won't set
apexd.status='ready' property if ro.apex.updatable is false, as it
should be set by the "main" invocation of apexd.

Test: atest apexservice_test
Change-Id: I44baa1942506d7fac8eef0148176f9f2f5989b31