Return preinstalled path with package information.
PackageManager is going to use the information about the original
partition in which a given apex package is preinstalled to determine
what capabilities should be given to the APKs contained in said apex.

Refactor the code to bundle together other preinstalled data that we
already scan such as the signing keys.

Bug: 138429615
Test: atest --test-mapping system/apex/apexd
Change-Id: I228de6b804f4512604b5a4dc44f349d306abd66d