Rebrand the term "rollback" to "revert" in apexd
In staged install flow, "rollback" usually means installation of a
version downgrade by RollbackManager. In apexd, the term means
bringing back the active sessions in /data/apex/active to the way it
was before the device was rebooted.

Using different terms for the two concepts in staged install flow
makes things easier to understand.

Bug: 144831762
Test: Builds without error
Change-Id: Ia648a82f09cd50aba5b0ab4d9e902120a612514e