Add ApexSessionParams to allow rollback details to be included when submitting a session.
This adds params to submitStagedSession to allow apexd to be informed
when a session has rollback enabled, or is a rollback, and to allow
the rollback id to be provided.

This information will be used to allow the apex data directories to
be snapshotted or restored (in a future CL).

This is an alternate approach to ag/9826517.

Bug: 141148175
Test: atest CtsRollbackManagerHostTestCases
Test: atest CtsStagedInstallHostTestCases
Change-Id: I6e9f16a4e66392d110fc13c9299a498679d01521
Merged-In: I6e9f16a4e66392d110fc13c9299a498679d01521