Do not exit(1) even if system_ext/apex doesn't exist
For now, apexd-bootstrap exits with 1 since most targets don't have
system_ext/apex which is one of bootstrap apex locations.
In practice, this is not a problem since scanning system_ext/apex is the
last action of apexd-bootstrap and init process(parent of
apexd-bootstrap) doesn't check if apexd-bootstrap succeeds or not.

This change tries to make apexd-bootstrap exit successfuly even if
system_ext/apex doesn't exist.

Bug: 146589762
Test: check logcat while booting
Change-Id: I84156005aef291e81636adf62b2e3ec4bd3a858f