Add hashtree_loop_name to MountedApexData
And also fill in hashtree_loop_name inside
MountedApexDatabase::PopulateFromMounts.

This is a prerequisite for freeing a loop device when an apex is
deactivated (e.g. during userspace reboot). Actual logic to free a loop
device will be implemented in a follow up CL.

Since I was there, added tests for
MountedApexDatabase::PopulateFromMounts function.

Test: atest apex_database_test
Test: atest apexservice_test
Test: atest --test-mapping system/apex:presubmit
Bug: 135984674
Bug: 145670581
Change-Id: Ib2a4ef30073f26931d7e2e326b589b929166e223