Restore DE_sys snapshot when rollback APEX is activated.
This replaces the contents of the DE_sys apex data directory
(/data/misc/apexdata/<apex name>) with the snapshot, when an
APEX session that is a rollback is activated.

Bug: 141148175
Test: Add dummy files to data directory for an apex and install that apex with
rollback enabled. Delete files in data directory and add new files. Trigger
rollback and reboot. Check that original files are restored. (Tested on internal)

Change-Id: I6556f7c7c1846e296550443454e767932ef49853