Store native crashing process in APEX session state
In the case of native crashes during boot, store the
updatable_crashing_process_name property in the APEX
session state. Note that this information will not
currently persist after the boot due to FS checkpointing. The
information will be used once sessions are stored in /metadata,
which will happen in a follow-up CL.

Bug: 138782888
Test: atest apexservice_test
Test: atest apex_rollback_tests, add logs to check the property
      is being set to the correct value

Change-Id: Ic38afd421ee50a0ac3100b17f6c764ce58bc87c7