Add a signing_args in apexer
In the build signing environment, we don't necessary have access to the
physical private key. And we work around it by passing an extra signing_helper
to the avbtool instead of a private key. Since we need to perform signing
to add the hashtree footer when rebuilding the apex image; apexer should recognize
the signing_helper and pass it to avbtool.

Bug: 148627666
Test: Call apexer with the argument
Change-Id: Ibf75d629d369bdbf4e3305b59eb495ccbfdc6428