Set apexd.status to "activated" between activation and DE snapshot.
- Cherry-pick of aosp/1228581

As of aosp/1224611, there is a new step in the preparation of APEXes
where init calls back into apexd after DE user data is unencrypted to
allow DE apex data to be snapshotted or restored.

Until this occurs, APEXes should not be used, so apexd.status should
not be set to "ready".

This CL introduces a new status of "activated" that init will wait on
before performing the configuration in the "perform_apex_config" step.

The status of "ready" will not be set only once any needed DE snapshot
or restore has been performed.

Bug: 148672144
Test: build & flash, check boot completes and check in logs that apexd
status was set correctly.
Merged-In: I3d840e414467f3c5934d158114c3cba2029988b9

Change-Id: If698214d92346157db0a4dfd27de02c6cccd7fe5