Account for fs checkpointing retry count now being 2
Due to the fs checkpointing retry count now being 2, changes
need to be made to apexd in to facilitate this. It is now
possible for an apex session to be in the ACTIVATED state, while
/data/apex has already been reverted due to checkpointing.
Therefore, scanStagedSessionsDirAndStage now needs to consider the
set of ACTIVATED sessions, as well as the STAGED ones.

Add/update tests that test the checkpointing functionality.

Bug: 148270009
Test: atest ApexRollbackTests
Change-Id: Ibe39bdf1513ee26a88e97cbf5c5340080ff95db6