Add ApexService method to delete CE snapshots.
This adds a new ApexService method to allow CE snapshots to be deleted.

This method is deleteCESnapshotsNotSpecified, as it deletes all
CE snapshots for the specified user, except for those will rollback ids
in a provided list.

RollbackManager will call this method when the user unlocks, passing a
list of all current rollbacks. This will ensure that all expired
snapshots are deleted. It will also allow any "orphaned" snapshots
(e.g. from tests or previous snapshots that have not been deleted) to
be cleaned up.

The new method is expected to be called when the user has unlocked -
it will fail if the user is not unlocked.

Bug: 147806409
Test: atest ApexServiceTest#DestroyCeSnapshotsNotSpecified
Change-Id: Ib83ccd3209947a10da9bf070dbda7e1f16993665