Make init.userspace_reboot.is_supported a rw property
General recommendation is to avoid read-only properties, and instead
control "read-onlines" by only allowing init/vendor_init to set the
property.

Since ro.init.userspace_reboot.is_supported was added in this release,
and nobody outside of the platform is querying it directly, it should
be fine to simply rename it.

Bug: 152803929
Test: adb shell getprop init.userspace_reboot.is_supported
Test: atest --test-mapping system/apex/tests
Change-Id: I3733c7ce3d68fdfafbcca878724988d65ccf6bea