apexd: remove two most spammy log lines
"Found bundled key in package" is the biggest contributor, while not
providing a lot of value for debugging (we are expecting all the APEXes
to have bundled key anyway).

"Skipped when bootstrapping" doesn't provide a lot of information as
well. We are more interested in the APEXes that weren't skipped during
bootstrap, and we already log them.

Test: adb reboot, checked logcat
Bug: 151419635
Change-Id: Ic9c99f47309aabbdb1c4f24a5f387ad609f06139