Create a new apex file for CTS
Create an Apex that it has the same module name with others, but the
package name is different. Use the apex to test the scenario: When
the system installs a Apex file, the module name of the Apex is known
by system, but the package name isn't, it can't be installed.

Test: m -j com.android.apex.cts.shim.v2_different_package_name
Bug: 154025817
Change-Id: I9d9e9615670758d1007a1280da7f4b0a6735f4c2