apexd: Add remountPackages
This can be used to speed up development cycle for module authors. Here
is an example of how this can be used:

1. adb shell stop
2. adb sync
3. adb shell cmd -w apexservice remountPackages
4. adb shell start

This won't help all the modules (e.g. bootstap apexes or adbd apex
unfortunately can't use this command), but should help modules that only
depend on the framework.

Test: atest ApexTestCases
Test: atest apexd_host_tests
Bug: 155275196
Change-Id: I32fe96385ab0494981c3f4c1fdc7d9c0bcd37b55