BACKPORT: Keep polling for loop device until ueventd cold boot is done
apexd-bootstrap runs in parallel with ueventd in order to optimize boot
time. In rare cases this might result in apexd-bootstrap trying to mount
an apex before ueventd created a loop device for it. This will result in
apexd-bootstrap failing to finish successfully and rebooting device into
bootloader.

To work around this issue, apexd now keeps polling for loop until
ueventd finishes it's cold boot sequence.

Test: add usleep(1s) to ueventd and verified device booted
Test: atest ApexTestCases
Bug: 169092045
Change-Id: Ibbfec780fc74a2956d6dd0b3bcd0c3fae4f1f6f5