egl: android: add IMapper@4 metadata API buffer_info getter
Starting from Android-11 Google introduces generalized API
to access buffer information. This API is a part of IMapper@4 HAL.

Implementation of the getter require additional libraries, which is
not available in Android NDK today, making using this getter hard
or impossible for projects that uses meson build system. Therefore
enable it only for in-AOSP builds.

Signed-off-by: Roman Stratiienko <r.stratiienko@gmail.com>