dt-bindings: arm: sunxi: Expand MBUS binding
The MBUS provides more than address translation and bandwidth control.
It provides a PMU to measure bandwidth usage by certain masters, and
provide notification via IRQ when they are active or idle.

The MBUS is also tightly integrated with the DRAM controller to provide
a Memory Dynamic Frequency Scaling (MDFS) feature. In view of this, the
MBUS binding needs to represent the hardware resources needed for MDFS,
which include the clocks and MMIO range of the adjacent DRAM controller.

Signed-off-by: Samuel Holland <samuel@sholland.org>