drm_hwcomposer: Quiet noisy errors when planes don't support various attributes
If a plane doesn't support alpha or rotation, and IsValidForLayer()
fails, we see lots of logcat noise to the effect of:
08-24 04:45:42.957   453   453 E hwc-platform: Failed to emplace layer 0, dropping it
08-24 04:45:42.957   453   453 E hwc-platform: Failed provision stage with ret -22
08-24 04:45:42.957   453   453 E hwc-drm-display-composition: Planner failed provisioning planes ret=-22
08-24 04:45:42.957   453   453 E hwc-drm-two: Failed to plan the composition ret=-22

This noise is unneccessarily worrisome, as they don't really
help explain why things fail, and we still fall back to client
composing and frames are correctly composited.

Thus, this patch switches the errors to verbose-level warnings,
which match the level in IsValidForLayer() which explain what
actually is going wrong.

Change-Id: I7ed06906b8d9e02e6ec0ac50a94346e9f9c05ac6
Signed-off-by: John Stultz <john.stultz@linaro.org>